homework_examples/week03-homework/pyproject.toml
[project]
name = "week03-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"common",
	"llama-index>=0.14.2",
	"llama-index-core>=0.14.2",
	"llama-index-embeddings-dashscope>=0.4.1",
	"llama-index-embeddings-huggingface>=0.6.1",
	"llama-index-embeddings-openai>=0.5.1",
	"llama-index-llms-dashscope>=0.5.1",
	"llama-index-llms-openai>=0.5.6",
	"llama-index-llms-openai-like>=0.5.1",
	"openai>=1.107.3",
	"paddle",
	"paddleocr<3.0",
	"paddlepaddle<=2.6",
	"pandas>=2.2.3",
	"python-dotenv",
	# add your dependencies here and run `uv sync` to install them
	"tabulate>=0.9.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
